Changed bindFields to work based on hasClass
bindFields now runs on a per-input basis, rather than globally. Once a field is bound, it is given a class 'ea-bound' to show that it is bound. Inputs with this class will never get re-bound (unless force == true). This allows for dynamic creation of dom elements on the page and then calling window.ea.reInit(); to setup the expressive annotations validation which will add the 'ea-bound' class to the newly created inputs.